requirements.txt
aiohttp==3.9.3
aiohttp-socks==0.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-property==0.2.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.19.1
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
curl_cffi==0.6.2
defusedxml==0.7.1
duckduckgo_search==5.1.0
fastapi==0.110.0
Flask==3.0.2
frozenlist==1.4.1
g4f==0.3.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.3
kaitaistruct==0.10
logtail-python==0.2.10
loguru==0.7.2
lxml==5.1.0
lz4==4.3.3
MarkupSafe==2.1.5
mmh3==4.1.0
msgpack==1.0.8
multidict==6.0.5
nest-asyncio==1.6.0
outcome==1.3.0.post0
phonenumbers==8.13.36
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.36
py-arkose-generator==0.0.0.2
pyasn1==0.5.1
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pydivert==2.1.0
PyExecJS==1.5.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-socks==2.4.4
questionary==2.0.1
requests==2.31.0
selenium==4.18.1
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.36.3
tinycss2==1.2.1
trio==0.25.0
trio-websocket==0.11.1
truecallerpy==1.0.3
trunofficial==0.2.4
typing_extensions==4.10.0
undetected-chromedriver==3.5.5
urllib3==2.2.1
uvicorn==0.28.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.4
zstandard==0.22.0
